+++ library/partials/header-listing.php
@@ -31,10 +31,30 @@
 			</div>
 
 			<div class="m-1of2 t-1of3 d-1of6 left pad-top">
-				<h4>Square Footage</h4>
-					<?php echo(types_render_field( "square-footage", array( 'raw' => true) )); ?> SF
-				<h4>Lease Price</h4>
-					$<?php echo(types_render_field( "lease-rate", array( 'raw' => true) )); ?>.00/SF
+				<?php if ( has_term( 'retail' , 'specialty') && has_term( 'office' , 'service') ) { ?>
+						<div class="d-1of2 t-1of2 m-1of2">
+							<?php if ( types_render_field( "square-footage" ) != null ) { ?>
+								<h4>Square Footage</h4>
+								<?php echo types_render_field( "square-footage", array( 'raw' => true) ); ?>
+							<?php } ?>
+						</div>
+					<?php } ?>
+
+					<?php if ( has_term( 'multifamily' , 'specialty') ) { ?>
+						<div class="d-1of2 t-1of2 m-1of2">
+							<?php if ( types_render_field( "units" ) != null ) { ?>
+								<h4>Units</h4>
+								<?php echo types_render_field( "units", array( 'raw' => true) ); ?>
+							<?php } ?>
+						</div>
+					<?php } ?>
+
+					<div class="d-1of2 t-1of2 m-1of2">
+						<?php if ( types_render_field( "lease-rate" ) != null ) { ?>
+							<h4>Lease Price</h4>
+							<?php echo types_render_field( "lease-rate", array( 'raw' => true) ); ?>
+						<?php } ?>
+					</div>
 			</div>
 
 			<a class="m-1of2 download-icon pull-l-1of12" href="<?php echo types_render_field( "report-pdf", array( 'raw' => true)); ?>"></a>