+++ post-formats/content-listing.php
@@ -16,19 +16,31 @@
 				<div class="tag red">
 					<?php // echo get_the_term_list( '', 'listing-type', '', '', ''); ?>
 				</div> -->
-
+				
 				<div class="listing-post-details cf">
-					<div class="d-1of2 t-1of2 m-1of2">
-					<?php if ( types_render_field( "square-footage" ) != null ) { ?>
-						<h4>Square Footage</h4>
-						<?php echo(types_render_field( "square-footage", array( 'raw' => true) )); ?> SF
+					
+					<?php if ( has_term( 'retail' , 'specialty') && has_term( 'office' , 'service') ) { ?>
+						<div class="d-1of2 t-1of2 m-1of2">
+							<?php if ( types_render_field( "square-footage" ) != null ) { ?>
+								<h4>Square Footage</h4>
+								<?php echo types_render_field( "square-footage", array( 'raw' => true) ); ?>
+							<?php } ?>
+						</div>
 					<?php } ?>
-					</div>
+
+					<?php if ( has_term( 'multifamily' , 'specialty') ) { ?>
+						<div class="d-1of2 t-1of2 m-1of2">
+							<?php if ( types_render_field( "units" ) != null ) { ?>
+								<h4>Units</h4>
+								<?php echo types_render_field( "units", array( 'raw' => true) ); ?>
+							<?php } ?>
+						</div>
+					<?php } ?>
 
 					<div class="d-1of2 t-1of2 m-1of2">
 						<?php if ( types_render_field( "lease-rate" ) != null ) { ?>
 							<h4>Lease Price</h4>
-							$<?php echo(types_render_field( "lease-rate", array( 'raw' => true) )); ?>.00/SF
+							<?php echo types_render_field( "lease-rate", array( 'raw' => true) ); ?>
 						<?php } ?>
 					</div>
 				</div>