+++ CHANGELOG.md
@@ -7,6 +7,19 @@ Author: Patrick Hanus
 
 *******************************************************************
 
+/* 1.03 */
+-comment out search form (For now)
+-Added Snazzy Maps Bower components
+-Added Snazzy maps components in header
+-Added conditional code for listing header to show SF or Units
+-Increased size of the 
+-Fixed the pending section on the Wernke’s Website
+-Increased font size on h4 for listings
+-Added a new post format for closed listings
+-Fixed broker’s phone -> cell phone
+-fixed the format of client logos for broker profiles
+-added a class for “.cta-underline-gray”
+
 /* 1.02 */
 - Added Bing Webmaster tools verification
 - Changed some default http://schema.org/ declarations